Enable metrics on windows
